feat: migrate server→client broadcasts from msgpack to protobuf
Converts all high-frequency game server broadcasts (state_delta, despawn,
world_snapshot, slot_spin, slot_counter, ability, coin_spawn, queue_update)
to protobuf encoding via a GameMessage oneof wrapper. Low-frequency Go
backend messages (pong, welcome, heat_update, reward) remain msgpack.

Client uses first-byte detection to route protobuf vs msgpack messages.
Removes Math.fround() since protobuf handles float32 natively.
Includes 16 roundtrip/size/format-detection tests and Docker build fixes.

+++ game/server/src/game/GameLoop.ts
@@ -203,19 +203,19 @@ export class GameLoop {
         [rot.x, rot.y, rot.z, rot.w]
       );
 
-      // Add to updates — Math.fround forces Float32 encoding in msgpack (5 bytes vs 9 bytes)
+      // Quantize to 3 decimal places; protobuf float fields handle float32 natively
       updates.push({
         id,
         pos: [
-          Math.fround(Math.round(pos.x * f) / f),
-          Math.fround(Math.round(pos.y * f) / f),
-          Math.fround(Math.round(pos.z * f) / f),
+          Math.round(pos.x * f) / f,
+          Math.round(pos.y * f) / f,
+          Math.round(pos.z * f) / f,
         ],
         rot: [
-          Math.fround(Math.round(rot.x * f) / f),
-          Math.fround(Math.round(rot.y * f) / f),
-          Math.fround(Math.round(rot.z * f) / f),
-          Math.fround(Math.round(rot.w * f) / f),
+          Math.round(rot.x * f) / f,
+          Math.round(rot.y * f) / f,
+          Math.round(rot.z * f) / f,
+          Math.round(rot.w * f) / f,
         ],
       });
     });
@@ -253,7 +253,7 @@ export class GameLoop {
     const tAfterDespawn = performance.now();
 
     // 6. Update pusher z in game state
-    const pusherZ = Math.fround(Math.round(this.pusher.getCurrentZ() * f) / f);
+    const pusherZ = Math.round(this.pusher.getCurrentZ() * f) / f;
     this.gameState.updatePusherZ(pusherZ);
 
     this.tickCount++;
